Mike task5

+++ src/application/java/models/Word.java
@@ -13,8 +13,9 @@ public class Word {
 	public final static int SKIP = 4;
 	
 	String word;
-	double score;
+	int score;
 	String result;
+
 	
 	/** constructor
 	 * 
@@ -22,7 +23,8 @@ public class Word {
 	 * @param score
 	 */
 	
-	public Word(String word, double score, int result) {
+
+	public Word(String word, int score, int result) {
 		this.word = word;
 		this.score = score;
 		switch(result) {
@@ -56,7 +58,7 @@ public class Word {
 	 * @return score
 	 */
 	
-	public double getScore() {
+	public int getScore() {
 		return this.score;
 	}
